update user, category....

+++ controllers/categories.controller.js
@@ -1,4 +1,5 @@
 const Category = require('../models/categories.model');
+const Product = require('../models/products.model');
 
 module.exports.getCategories = async (req, res) => {
     try {
@@ -12,9 +13,9 @@ module.exports.getCategories = async (req, res) => {
 module.exports.createCategory = async (req, res) => {
     try {
         const { name } = req.body;
-        const category = await Category.findOne({ name });
+        const category = await Category.findOne({ name: name.toLowerCase() });
         if (category) return res.status(400).json({ msg: "this category already exists" });
-        const newCategory = new Category({ name });
+        const newCategory = new Category({ name: name.toLowerCase() });
         await newCategory.save();
         return res.json({ msg: "created a category", newCategory });
     } catch (error) {
@@ -23,6 +24,13 @@ module.exports.createCategory = async (req, res) => {
 }
 module.exports.deleteCategory = async (req, res) => {
     try {
+        const nameCategory = await Category.findOne({ _id: req.params.id })
+        const product = await Product.findOne({ category: nameCategory.name })
+        if (product) {
+            return res.status(400).json({
+                msg: "Please delete all products of this category"
+            })
+        }
         await Category.findByIdAndDelete(req.params.id);
         return res.json({ msg: "deleted a category" });
     } catch (error) {
